requirements.txt
biopython>=1.81
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
tmtools>=0.1
pandas>=2.0
click>=8.1
joblib>=1.3
